requirements.txt
PySide6
apscheduler
beautifulsoup4
lxml
requests
readability-lxml
feedparser
pyrogram
tgcrypto
telethon
yt-dlp
playwright
natasha
ollama
faster-whisper
torch
paddleocr
PyMuPDF
